code to create error and language models
frequencycounter can be run on irish or english to create the frequency files needed for the error model and the language model. Saves it using pickle so that the data is already in usable form.

combinedmodels is an adaptation of the bayesian model but uses the combination of error and language models. The file names are currently hardcoded in. might be worth exploring how to change the files names for multiple languages