Update views.py
I added a possible2 function for edit distance 2.  From there I used a short circuit technique to speed up the code.  The assumption I made with the short circuit technique is that misspellings  generated from fewer edits are more likely than misspelling generated with more edits.  So the code now checks the list of 1 edit distance possibilities, if nothing is there it then checks the list of 2 edit distance possibilities, if nothing there it gives back the original word.