Update index.html
Fixing a bug I found.  The highlight misspelled text wasn't working because of JSON.stringify.  I also found a pretty nasty bug involving the div id for highlightTextarea.  Essentially this j.query package creates a transparent div on top of a transparent div.  I had to drill down to get the correct div id, which for some reason the package names the exact same as the upper div.  I also added more functionality to the clear button.  It deletes the highlighted areas as well as just the text.