Update views_irish.py
The same functionality was added to the irish_views page as the english views page, two new functions to handle the back end of the context sensitive right click event on the front end.  get_context_words grabs the preceding two words of the word in the right click event.  clean_context_words edits the list of preceding words and deletes it if it is a misspelling.  Edits to the spell_check function were also made to read in the preceding two words into the corrections function.